+++ src/util/openDialog.ts
@@ -1,10 +1,17 @@
-export function openDialog(id: string, params?: unknown) {
+export function openDialog(
+    id: string,
+    params?: Partial<Record<string, unknown> & { lazy?: boolean }>,
+) {
+    const open = () => {
+        document.querySelector<HTMLDialogElement>(`#${id}`)
+            ?.showModal();
+    };
+
     if (params) {
         window.dispatchEvent(
-            new CustomEvent("dialog-open", { detail: { id, params } }),
+            new CustomEvent("dialog-open", { detail: { id, params, open } }),
         );
     }
 
-    document.querySelector<HTMLDialogElement>(`#${id}`)
-        ?.showModal();
+    if (!params?.lazy) open();
 }